requirements.txt
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
cryptography>=41.0.0
pyotp>=2.8.0
PyJWT>=2.8.0
fastapi>=0.100.0
starlette>=0.27.0
aiosmtplib>=2.0.0
dkimpy>=1.1.0
slowapi>=0.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
future~=1.0.0
pygetwindow~=0.0.9
opencv-python~=4.11.0.86
dotenv~=0.9.9
python-dotenv~=1.0.1
uvicorn~=0.35.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
bcrypt>=4.0.0
python-jose>=3.3.0
Authlib>=1.0.0
aiosqlite>=0.19.0
